+++ ChangeLog.md
@@ -5,6 +5,4 @@
 - Starting rewrite of the gpdviz system.
 - Using [scala-tsi](https://github.com/scala-tsi/scala-tsi)
   to help capture the data models from backend to frontend.
-- note that a rewrite of the frontend has alreadu been started
-  in a separate repo but will eventually be moved here.
-  
+- move the initial rewrite of the frontend to this repo.